requirements.txt
selenium==4.1.0
pytest==6.2.5
allure-pytest==2.9.45
